Replace Language ADT, use Dict internally
The type safety we get from using an ADT doesn't apply to how we
use it. Now it's easier to add languages without a major verion bump.

Since we're a adding a bunch of languages, having a record field for
each language starts to add up to our runtime memory footprint.
Using a Dict _and_ filtering empty values might actually be an
improvement over the record variant.